requirements.txt
confluent-kafka
pycryptodome
PyYAML